Enable legend toggle
